refactor: move `lsp` to `ide-lsp`
