+++ ide-lsp/src/main/shell/aya-lsp.bat
