+++ ide-lsp/src/main/shell/aya.bat
