+++ ide-lsp/src/test/resources/lsp-test-lib/aya.json
