+++ ide-lsp/src/test/resources/lsp-test-lib/src/Nat.aya
