+++ ide-lsp/src/test/resources/lsp-test-lib/src/Vec.aya
