+++ settings.gradle.kts
@@ -1,4 +1,4 @@
-// Copyright (c) 2020-2022 Tesla (Yinsen) Zhang.
+// Copyright (c) 2020-2023 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
 
 rootProject.name = "aya-prover"
@@ -17,5 +17,5 @@ include(
   "pretty",
   "parser",
   "ide",
-  "lsp",
+  "ide-lsp",
 )
